% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adpcr2panel.R
\name{adpcr2panel}
\alias{adpcr2panel}
\title{Convert adpcr object to array}
\usage{
adpcr2panel(input, use_breaks = FALSE)
}
\arguments{
\item{input}{object of the \code{\linkS4class{adpcr}} class.}

\item{use_breaks}{if \code{TRUE}, input is cutted into intervals using the
\code{breaks} slot. If \code{FALSE}, input is converted to factor using
\code{\link[base]{as.factor}}. Ignored if data has \code{"np"} type (see
possible types of \code{\linkS4class{adpcr}} objects).}
}
\value{
A named list of length equal to the number of arrays in the \code{input}.
Each element is a single array in matrix-like form, where dimensions are set
exactly as in case of the real plate. Names of the list corresponds to the names
of assays (\code{tnp} data) or runs (any other type of \code{\linkS4class{adpcr}}
data).

The matrices contain values from array, either integers (when \code{use_break} is
\code{FALSE}) or characters (when \code{use_break} is \code{TRUE}).
}
\description{
\code{\linkS4class{adpcr}} objects contains all data from array-based digital
PCR experiments. This function converts them into matrices, where rows and
columns represent rows and columns in a real array.
}
\author{
Michal Burdukiewicz.
}
\seealso{
\code{\link{extract_dpcr}}.
}

