% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioradCNV.R
\docType{data}
\name{BioradCNV}
\alias{BioradCNV}
\title{Copy number variation experiment}
\format{An object of class \code{\linkS4class{adpcr}} with \code{"tnp"} type containing 
four runs from seven experiments (four runs per each experiment).}
\source{
Dorazio RM, Hunter ME, \emph{Statistical Models for the Analysis 
and Design of Digital Polymerase Chain Reaction (dPCR) Experiments}. 
Analytical Chemistry 2015. 87(21): p.10886-10893
}
\description{
Copy Number Variation of MYC Gene for seven patients measured using QX200 (Bio-Rad)
}
\examples{
data(BioradCNV)
summary(BioradCNV)
}
\author{
Robert M. Dorazio, Margaret E. Hunter.
}
