% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coordinates.R
\name{calc_coordinates}
\alias{calc_coordinates}
\title{Calculate Array Coordinates}
\usage{
calc_coordinates(array, half)
}
\arguments{
\item{array}{A single array, as generated by \code{\link{adpcr2panel}}.}

\item{half}{If \code{left} or \code{right}, every well is represented only
by the adequate half of the rectangle.}
}
\value{
Returns two sets of coordinates of each microfluidic well:
\code{coords} is a list of coordinates suitable for usage with functions from
\code{\link{graphics}} package. The second element is a data frame of coordinates 
useful for users utilizing ggplot2 package.
}
\description{
Calculates coordinates of points on plot to represent a digital PCR array.
}
\seealso{
\code{\link{plot_panel}} - plots \code{\linkS4class{adpcr}} data.
\code{\link{adpcr2panel}} - converts \code{\linkS4class{adpcr}} object to arrays.
}
\author{
Michal Burdukiewicz, Stefan Roediger.
}
\keyword{manip}
