% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2dpcr.R
\name{df2dpcr}
\alias{df2dpcr}
\title{Convert data.frame to dpcr object}
\usage{
df2dpcr(df)
}
\value{
An object of \code{\linkS4class{adpcr}} or \code{\linkS4class{ddpcr}} type, 
depends on the presence of additional three columns with names of rows and columns of 
panels as well as panel indices (see Details).
}
\description{
Converts \code{\link{data.frame}} object to to \code{\linkS4class{adpcr}} or 
\code{\linkS4class{ddpcr}} object. The resulting object will have \code{"tnp"}
type.
}
\details{
The data frame must have following columns with exactly specified names:
\describe{
\item{experiment}{names of experiments}
\item{replicate}{indices of replicates}
\item{assay}{names of assays}
\item{k}{number of positive partitions}
\item{n}{total number of partitions}
}
There are also three optional columns:
\describe{
\item{row_names}{names of rows}
\item{col_names}{names of columns}
\item{panel_id}{indices of panels}
}

If these additional columns are present, the resulting object has 
\code{\linkS4class{adpcr}} type.
}
\examples{
dat <- data.frame(experiment = factor(rep(paste0("Experiment", 1L:2), 3)),
                  replicate = c(1, 1, 2, 2, 3, 3),
                  assay = "Assay1",
                  k = c(55, 121, 43, 150, 70, 131),
                 n = 765)
df2dpcr(dat)
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\seealso{
Flexibly create \code{\linkS4class{dpcr}} objects: \code{\link{create_dpcr}}
}
\keyword{utilities}

