% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dpcr-class}
\alias{dpcr}
\alias{dpcr-class}
\title{Class \code{"dpcr"} - general digital PCR}
\description{
A class containing results of any digital PCR experiment.
Type of data in all columns is specified in slot \code{"type"}.
}
\details{
Possible \code{type} values of \code{dpcr} objects:
\enumerate{
 \item{\code{"ct"}: cycle threshold of each partition,}
 \item{\code{"fluo"}: fluorescence of each partition,}
 \item{\code{"nm"}: number of molecules in each partition,}
 \item{\code{"np"}: status (positive (1) or negative(0)) of each partition,}
 \item{\code{"tnp"}: total number of positive partitions in the run (single
 value per each run, not per partition).}}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\code{"matrix"} containing data from array. See Description.}

\item{\code{n}}{\code{"integer"} equal to the number of partitions in each
experiment.}

\item{\code{exper}}{\code{"factor"} representing the id or name of experiments.}

\item{\code{replicate}}{\code{"factor"} representing the id or name of replicate.}

\item{\code{assay}}{\code{"factor"} representing the id or name of the assay.}

\item{\code{type}}{Object of class \code{"character"} defining type of data. See Details.}
}}
\note{
This class should not be directly used. Instead, users should use more
specific class: \code{\linkS4class{adpcr}}, \code{\linkS4class{ddpcr}} or
\code{\linkS4class{rtadpcr}}
}
\author{
Michal Burdukiewicz.
}
\keyword{classes}

