% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcr2df.R
\name{dpcr2df-methods}
\alias{dpcr2df-methods}
\alias{dpcr2df}
\alias{dpcr2df,adpcr-method}
\alias{dpcr2df,dpcr-method}
\title{Convert dpcr object to data frame}
\usage{
dpcr2df(input)
}
\arguments{
\item{input}{\code{\linkS4class{adpcr}} or \code{\linkS4class{dpcr}} object.}
}
\value{
data frame with 5 (if input was  \code{\linkS4class{dpcr}}) or 
8 columns (if input was  \code{\linkS4class{adpcr}}).
}
\description{
Converts \code{\linkS4class{adpcr}} or \code{\linkS4class{dpcr}} object to 
\code{\link{data.frame}} object.
}
\examples{
dpcr2df(six_panels)
}
\seealso{
Inverse function: \code{\link{df2dpcr}}
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\keyword{utilities}
