% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpcr_density_gui}
\alias{dpcr_density_gui}
\title{Digital PCR Density Graphical User Interface}
\usage{
dpcr_density_gui()
}
\description{
Launches graphical user interface that allows calculating density of
positive partitions distribution.
}
\section{Warning }{
 Any ad-blocking software may be cause of malfunctions.
}
\author{
Michal Burdukiewicz, Stefan Roediger.
}
\seealso{
\code{\link{dpcr_density}}.
}
\keyword{hplot}

