% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcReport_gui.R
\name{dpcReport}
\alias{dpcReport}
\title{Digital PCR Report Graphical User Interface}
\usage{
dpcReport()
}
\description{
Launches graphical user interface that generates reports from digital PCR 
data.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\author{
Michal Burdukiewicz, Stefan Roediger.
}
\keyword{hplot}
