% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRmethyl.R
\docType{data}
\name{dPCRmethyl}
\alias{dPCRmethyl}
\title{Methylated human gDNA}
\format{A \code{\linkS4class{adpcr}} object}
\source{
Mario Menschikowski group /  Institute of Clinical Chemistry and 
Laboratory Medicine, TU-Dresden, Dresden, Germany
}
\description{
Results of dPCR experiment: serial dilutions (0.25. 0.50, 0.75, 1.00) of 
methylated human gDNA. For each concentration level, three samples were measured 
using QX100 (Bio-Rad).
}
\examples{
summary(dPCRmethyl)

}
\author{
Mario Menschikowski, Stefan Roediger, Michal Burdukiewcz
}
\keyword{datasets}
