% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_poisson.R
\name{is_poisson}
\alias{is_poisson}
\title{Check if Sample Comes from Poisson Distribution}
\usage{
is_poisson(k, n, times = 10000)
}
\arguments{
\item{x}{A vector of counts, \code{\linkS4class{adpcr}}, \code{\linkS4class{ddpcr}} or
\code{\linkS4class{qdpcr}} object.}
}
\value{
An object of \code{is_poisson_c} class.
}
\description{
Calculates coordinates of points on plot to represent a digital PCR array.
}
\author{
Michal Burdukiewicz, Piotr Sobczyk, Stefan Roediger.
}
\keyword{htest}

