% Generated by roxygen2 (4.0.1): do not edit by hand
\name{num2int}
\alias{num2int}
\title{Convert numeric to integer}
\usage{
num2int(x)
}
\description{
Converts \code{numeric} values to positive \code{integer}s with a warning.
}
\details{
num2int uses \code{\link[base]{as.integer}} functionality.
}
\examples{
num2int(pi) == 3L
}
\keyword{manip}

