% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdpcr.R
\docType{class}
\name{qdpcr-class}
\alias{qdpcr}
\alias{qdpcr-class}
\alias{show,qdpcr-method}
\alias{show.qdpcr}
\alias{summary,qdpcr-method}
\alias{summary.qdpcr}
\title{Class \code{"qdpcr"}}
\description{
An object representing digital PCR reaction depicted as Poisson process.
}
\section{Slots}{
 \describe{
\item{list(".Data")}{\code{"matrix"} with three
columns containing: number of cycles, amplification curves and cumulative
sum of events.}
\item{list("mu")}{\code{"numeric"} of the expected number of events in
defined interval.}
\item{list("CT")}{\code{"numeric"} value of the
"average time" between the occurrence of a positive reaction and another
positive reaction.}
\item{list("partitions")}{\code{"integer"} value equal to the number of
partitions.}
\item{list("events")}{\code{"integer"} value equal number of
events (positive partitions taken to further
analysis)}
}
}
\author{
Stefan Roediger, Michal Burdukiewicz.
}
\seealso{
\code{\link{plot.qdpcr}},
}
\keyword{classes}

