% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_amp}
\alias{read_amp}
\title{Read digital PCR amplitude raw data}
\usage{
read_amp(input, ext = NULL)
}
\arguments{
\item{input}{name of the input file (\code{character}) or input object 
(\code{data.frame}).}

\item{ext}{extension of the file ().}
}
\value{
An object of \code{\linkS4class{adpcr}}.
}
\description{
Reads digital PCR amplitude data.
}
\details{
The amplitude data means a compressed directory of 
amplification.
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\keyword{utilities}
