% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_BioMark}
\alias{read_BioMark}
\title{Read BioMark}
\usage{
read_BioMark(file)
}
\arguments{
\item{file}{name of the input file.}
}
\description{
Reads digital PCR data from the BioMark (Fluidigm) (Summary Table results)
}
\author{
Michal Burdukiewcz, Stefan Roediger
}

