% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_dpcr}
\alias{read_dpcr}
\title{Read digital PCR data}
\usage{
read_dpcr(file, format, ...)
}
\arguments{
\item{file}{name of the input file.}

\item{format}{of the file.}

\item{...}{additional arguments for the appropriate format.}
}
\description{
Reads digital PCR data in various formats.
}
\details{
Input files may be in .csv, .xls or .xlsx format. In case of Excel files with
multiple sheets, only the first sheet will be analyzed.
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\seealso{
\code{\link{read_raw}}
}

