% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_QX200}
\alias{read_QX200}
\title{Read QX200}
\usage{
read_QX200(input)
}
\arguments{
\item{input}{name of the input file (\code{character}) or input object 
(\code{data.frame}).}
}
\value{
An object of \code{\linkS4class{adpcr}} class.
}
\description{
Reads digital PCR data from the QX200 Droplet Digital PCR System (Bio-Rad).
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\seealso{
See \code{\link{read_dpcr}} for detailed description of input files.
}
\keyword{utilities}

