% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_raw}
\alias{read_raw}
\title{Read digital PCR raw data}
\usage{
read_raw(file)
}
\arguments{
\item{file}{name of the input file.}
}
\description{
Reads digital PCR data in raw format.
}
\author{
Michal Burdukiewcz, Stefan Roediger
}

