% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename object}
\usage{
rename(x, exper = NULL, replicate = NULL)
}
\arguments{
\item{x}{an \code{\linkS4class{adpcr}} or \code{\linkS4class{ddpcr}} object.}

\item{exper}{a \code{factor} of new experiments' names. If \code{NULL},
experiments' names are not changed.}

\item{replicate}{a \code{factor} of new replicates' ids. If \code{NULL},
replicates' names are not changed.}
}
\description{
Renames objects of class \code{\linkS4class{adpcr}} or \code{\linkS4class{ddpcr}}.
}
\keyword{manip}

