% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_ddpcr_bkm.R
\name{sim_ddpcr_bkm}
\alias{sim_ddpcr_bkm}
\title{Simulate Droplet Digital PCR}
\usage{
sim_ddpcr_bkm(m, n = 20000L, mexp = TRUE, n_exp = 8L, type = "np",
  fluo_range = NULL, sddropc = 0, mudropr = 1, sddropr = 0,
  Pvar = TRUE, piperr = 0, dropsd = 0, falpos = 0, falneg = 0,
  rain = 0)
}
\arguments{
\item{m}{is either the concentration of templates molecules in the raw sample
(copies/microliter)
or the expected number of template molecules per droplet.
See \code{mexp} parameter for details
Must be a (vector of) positive integers.}

\item{n}{the expected number of droplets per experiment. Must be a positive integer.
Default 20000 based on the Bio-Rad ddPCR QX100 theoretical expected values}

\item{mexp}{If \code{TRUE}, m is the expected number of template molecules per droplet
If \code{FALSE}, m is the concentration of the raw sample
Default \code{TRUE} as in Jacobs et al.}

\item{n_exp}{the number of experiments that are simulated by the function for each given
\code{m}.
Default 8 for eight replicates for each given \code{m} as in Jacobs et al. 2014}

\item{type}{Object of class \code{"character"} defining type of data. Could
be \code{"tnp"} (total number of positive wells in the panel), \code{"fluo"}
(fluorescence) or \code{"np"} (status (positive (1) or negative(0)) of each droplet).}

\item{fluo_range}{defines expected space between two
consecutive measured droplets. Used only when parameter \code{type} has value
\code{"fluo"}. Values between 10-20 give nice results.}

\item{sddropc}{standard deviation of the number of droplets generated.
Must be a real number between 0 and \code{n} divided by 10.
Default 0 for constant number of droplets.}

\item{mudropr}{average proportion (between 0 and 1) of retained partitions.
Must be a real number between 0 and 1.
Default 1 for no loss.}

\item{sddropr}{relative standard deviation of the proportion of retained partitions.
Must be a real number, preferably close to 0.
Default 0 for a constant loss.}

\item{Pvar}{If \code{TRUE}, number of copies in constant volume follows P(c) distribution.
If \code{FALSE},  number of copies in constant volume is constant.
Default \code{TRUE} for the realistic Poisson model.}

\item{piperr}{coefficient of variation of the actual pipetted volume from the raw material.
Must be a positive real number, preferably close to 0 (0.1 = 10% is very large).
Default 0 for constant volume equal to the expected volume}

\item{dropsd}{relative variability of the droplet volume.
parameter sigma of a lognormal distribution with mu = 0.
Must be a positive real number, preferably close to 0.
Default 0 for constant droplet size.}

\item{falpos}{probability that a partition containing no copy gives a positive result.
Must be a real number between 0 and 1.
Default 0 for no false positives.
Only used if \code{fluo} is \code{NULL}.}

\item{falneg}{probability that a partition containing at least one copy gives a negative
result
Must be a real number between 0 and 1
Default 0 for no false negatives
Only used with \code{fluo} is \code{NULL}}

\item{rain}{parameter that defines how much inhibition is enforced on positive droplets.
Must be a real number between 0 and 1 with 0 no rain, 1 positive droplets follow same
distribution as negative droplets
Default 0 for no rain. Used only when \code{type} is \code{fluo}.}
}
\description{
Simulates results of a droplet digital PCR.
}
\details{
sim_ddpcr_bkm is based on the R code from Jacobs et al. (2014) (see references).
}
\examples{
#two concentration, each 3 repetitions
dat2_3 <- sim_ddpcr_bkm(c(0.5, 0.6), n_exp = 3, type = "tnp")
}
\references{
Jacobs B, Goetghebeur E, Clement L \emph{Impact of variance components on reliability of
absolute quantification using digital PCR} BMC Bioinformatics, 2014.
}

