% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test_counts_gui.R
\name{test_counts_gui}
\alias{test_counts_gui}
\title{Compare digital PCR runs - interactive presentation}
\usage{
test_counts_gui()
}
\description{
Launches graphical user interface allowing multiple comparisons of simulated digital
PCR reactions.
}
\section{Warning }{
 Any ad-blocking software may be cause of malfunctions.
}
\author{
Michal Burdukiewicz, Stefan Roediger.
}
\seealso{
\code{\link{test_counts}}.
}
\keyword{hplot}

