% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/White.R
\docType{data}
\name{White}
\alias{White}
\title{Digitalized Data from a Fluidigm Array}
\format{A data frame with some observations on some variables.
\describe{ Here is nothing to describe }}
\source{
Data were digitalized from the supplement material
"1471-2164-10-116-S1.pdf" by White et al. (2009) BMC Genomics
}
\description{
These are the results data from the \code{White} data as measured by the
UT digital PCR on Fluidigm 12.765 digital Array. The data were digtilized
from a supplementory figure "1471-2164-10-116-S1.pdf"
by White et al. (2009) BMC Genomics
}
\details{
Setup: Experimental details were described be White et al. (2009) BMC
Genomics. The digitalization of the figure was done with imageJ and the
"MicroArray Profile" plugin by Bob Dougherty (rpd@optinav.com) and
Wayne Rasband.

Annotation: TBD
}
\examples{
#str(White)
}
\author{
Stefan Roediger, Michal Burdukiewcz, White et al. (2009) BMC Genomics
}
\references{
White RA, Blainey PC, Fan HC, Quake SR. Digital PCR provides
sensitive and absolute calibration for high throughput sequencing.
BMC Genomics 2009;10:116. doi:10.1186/1471-2164-10-116.

Dougherty B, Rasband W. MicroArray Profile ImageJ Plugin n.d.
http://www.optinav.com/imagej.html (accessed August 20, 2015).
}
\keyword{datasets}

