% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dens.R
\name{compare_dens}
\alias{compare_dens}
\title{Plot and Compare Densities}
\usage{
compare_dens(input, moments = TRUE, ...)
}
\arguments{
\item{input}{object of class \code{\linkS4class{dpcr}} containing only one run.}

\item{moments}{logical, if \code{TRUE}, both theoretical and empirical
moments are printed on the plot.}

\item{\dots}{other arguments passed to the \code{plot} function.}
}
\description{
Plot empirical and theoretical density of the result of the digital PCR
experiment.
}
\examples{

adpcr_big <- sim_adpcr(m = 35, n = 40, times = 50, pos_sums = FALSE, n_panels = 1)
compare_dens(adpcr_big, moments = TRUE)

}
\author{
Michal Burdukiewcz.
}
\seealso{
\link{moments} is used to calculate moments of Poisson distribution.
}
\keyword{PCR}
\keyword{density}
\keyword{digital}
\keyword{empirical}

