% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{count_test}
\alias{count_test}
\alias{count_test-class}
\title{Class \code{"count_test"}}
\description{
A class
}
\section{Slots}{

\describe{
\item{group_coef}{\code{"data.frame"} containing experiments, groups to which they
belong and calculated values of rate.}
\item{test}{\code{"matrix"} containing result of t-test.} }
}
\author{
Michal Burdukiewicz.
}
\seealso{
Nothing yet.
}
\keyword{classes}

