% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{count_test}
\alias{count_test}
\alias{count_test-class}
\alias{plot,count_test-method}
\alias{show,count_test-method}
\alias{summary,count_test-method}
\title{Class \code{"count_test"}}
\usage{
\S4method{summary}{count_test}(object)

\S4method{show}{count_test}(object)

\S4method{plot}{count_test}(x, aggregate = FALSE)
}
\arguments{
\item{object}{of class \code{count_test}.}

\item{x}{object of class \code{count_test}.}

\item{aggregate}{logical, if \code{TRUE} experiments are aggregated according
to their group.}
}
\description{
A class for results of \code{\link{test_counts}} function.
}
\details{
In case of aggregated plot, mean confidence intervals for groups are presented
as dashed lines.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summary statistics of assigned groups.

\item \code{show}: Print both \code{group_coef} and \code{t_res}

\item \code{plot}: 
}}
\section{Slots}{

\describe{
\item{group_coef}{\code{"data.frame"} containing experiments, groups to which they
belong and calculated values of rate.}
\item{t_res}{\code{"matrix"} containing result of t-test.} }
}
\author{
Michal Burdukiewicz.
}
\seealso{
Nothing yet.
}
\keyword{classes}

