\name{dpcr_density_gui}
\alias{dpcr_density_gui}

\title{
Digital PCR Density Graphical User Interface
}
\description{
Launches graphical user interface that allows calculating density of positive partitions distribution.
}
\usage{
dpcr_density_gui()
}


\author{
Michal Burdukiewicz, Stefan Roediger.
}

\section{Warning }{
Any ad-blocking software may be cause of malfunctions. 
}

\seealso{
\code{\link{dpcr_density}}.
}

\keyword{ hplot }
