% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcr_density_gui.R
\name{dpcr_density_gui}
\alias{dpcr_density_gui}
\title{Digital PCR Density Graphical User Interface}
\usage{
dpcr_density_gui()
}
\description{
Launches graphical user interface that allows calculating density of
positive partitions distribution.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}
\author{
Michal Burdukiewicz, Stefan Roediger.
}
\seealso{
\code{\link{dpcr_density}}.
}
\keyword{hplot}

