% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_peaks.R
\docType{data}
\name{many_peaks}
\alias{many_peaks}
\title{Artificial data set with many peaks}
\format{A data frame with 493 observations on the following 2 variables.
\describe{ 
\item{t}{a numeric vector} 
\item{F}{fluorescence
value} }}
\description{
A set of data from an artificial droplet flow experiment. It contains
various kinds of peaks - positive peaks, negative peaks and peak-like noise.
}
\examples{

data(many_peaks)
plot(many_peaks, type = "b")

}
\keyword{datasets}
