% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num2int}
\alias{num2int}
\title{Convert numeric to integer}
\usage{
num2int(x)
}
\arguments{
\item{x}{an \code{numeric} vector}
}
\description{
Converts \code{numeric} values to positive \code{integer}s with a warning.
}
\details{
num2int uses \code{\link[base]{as.integer}} functionality.
}
\examples{
suppressWarnings(num2int(pi) == 3L)
}
\keyword{manip}

