% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdpcr.R
\docType{class}
\name{qdpcr-class}
\alias{qdpcr}
\alias{qdpcr-class}
\alias{show,qdpcr-method}
\alias{show.qdpcr}
\alias{summary,qdpcr-method}
\alias{summary.qdpcr}
\title{Class \code{"qdpcr"}}
\description{
An object representing digital PCR reaction depicted as Poisson process. Inherits 
from \code{\linkS4class{dpcr}}.
}
\section{Slots}{
 \describe{ 
\item{list("mu")}{\code{"numeric"} of the expected number of events in a
defined interval.}
\item{list("C0")}{\code{"numeric"} the occurrence in a defined interval.}
\item{list("CT")}{\code{"numeric"} value of the
"average time" between the occurrence of a positive reaction and another
positive reaction.}
}
}
\author{
Stefan Roediger, Michal Burdukiewicz.
}
\seealso{
\code{\link{plot.qdpcr}},
}
\keyword{classes}

