% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qpcrpp.R
\docType{class}
\name{qpcrpp-class}
\alias{qpcrpp}
\alias{qpcrpp-class}
\alias{show,qpcrpp-method}
\alias{show.qpcrpp}
\alias{summary,qpcrpp-method}
\alias{summary.qpcrpp}
\title{Class \code{"qpcrpp"}}
\description{
An object representing digital PCR reaction depicted as Poisson process.
}
\section{Slots}{
 \describe{ \item{list(".Data")}{\code{"matrix"} with three
columns containing: number of cycles, amplification curves and cumulative
sum of events.}\item{:}{\code{"matrix"} with three columns containing:
number of cycles, amplification curves and cumulative sum of events.}
\item{list("mu")}{\code{"numeric"} of the expected number of events in
defined interval.}\item{:}{\code{"numeric"} of the expected number of events
in defined interval.} \item{list("CT")}{\code{"numeric"} value of the
"average time" between the occurrence of a positive reaction and another
positive reaction.}\item{:}{\code{"numeric"} value of the "average time"
between the occurrence of a positive reaction and another positive reaction.}
\item{list("partitions")}{\code{"integer"} value equal to the number of
partitions.}\item{:}{\code{"integer"} value equal to the number of
partitions.} \item{list("events")}{\code{"integer"} value equal number of
events (positive partitions taken to further
analysis)}\item{:}{\code{"integer"} value equal number of events (positive
partitions taken to further analysis)} }
}
\author{
Stefan Roediger, Michal Burdukiewicz.
}
\seealso{
\code{\link{plot.qpcrpp}},
}
\keyword{classes}

