% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_QX100}
\alias{read_QX100}
\title{Read QX100}
\usage{
read_QX100(file)
}
\arguments{
\item{file}{name of the input file.}
}
\description{
Reads digital PCR data from the QX100 Droplet Digital PCR System (Bio-Rad).
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\seealso{
See \code{\link{read_dpcr}} for detailed description of input files.
}

