% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_raw}
\alias{read_raw}
\title{Read digital PCR raw data}
\usage{
read_raw(file, adpcr)
}
\arguments{
\item{file}{name of the input file.}

\item{adpcr}{\code{logical}. If \code{TRUE}, function creates
\code{\linkS4class{adpcr}} object. If \code{FALSE}, function creates
\code{\linkS4class{ddpcr}} object.}
}
\description{
Reads digital PCR data in raw format.
}
\author{
Michal Burdukiewcz, Stefan Roediger
}

