% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dpcr.R
\name{read_redf}
\alias{read_redf}
\title{Read digital PCR raw data}
\usage{
read_redf(input, ext = NULL)
}
\arguments{
\item{input}{name of the input file (\code{character}) or input object 
(\code{data.frame}).}

\item{ext}{extension of the file ().}
}
\value{
An object of \code{\linkS4class{adpcr}} or \code{\linkS4class{dpcr}} type, 
depends on the value of \code{adpcr} parameter.
}
\description{
Reads REDF (Raw Exchange Digital PCR format) data.
}
\author{
Michal Burdukiewcz, Stefan Roediger
}
\keyword{utilities}

