% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dpcr.R
\name{rename_dpcr}
\alias{rename_dpcr}
\title{Rename object}
\usage{
rename_dpcr(x, exper = NULL, replicate = NULL, assay = NULL)
}
\arguments{
\item{x}{an \code{\linkS4class{adpcr}} or \code{\linkS4class{dpcr}} object.}

\item{exper}{a vector of new experiments' names. If \code{NULL},
experiments' names are not changed.}

\item{replicate}{a vector of new replicates' ids. If \code{NULL},
replicates' names are not changed.}

\item{assay}{a vector of new assays' names. If \code{NULL},
assays' names are not changed.}
}
\description{
Renames objects of class \code{\linkS4class{adpcr}} or \code{\linkS4class{dpcr}}.
}
\details{
The valid \code{exper}, \code{replicate} and \code{assay} names 
are factors. For the sake of convenience, this function converts other types 
to factors if it is possible.
}
\keyword{manip}
