% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtadpcr.R
\docType{class}
\name{rtadpcr-class}
\alias{rtadpcr-class}
\alias{rtadpcr}
\title{Class \code{"rtadpcr"} - real-time array digital PCR experiments}
\description{
A class designed to contain results from real-time array digital PCR
experiments. Data is represented as matrix, where each column describes
different measurement point (i.e. cycle number) and every row different
partition.
}
\section{Slots}{
 \describe{ \item{list(".Data")}{\code{"matrix"} containing
data from array. See Description.}\item{:}{\code{"matrix"} containing data
from array. See Description.} \item{list("n")}{Object of class
\code{"integer"} equal to the number of partitions.}\item{:}{Object of class
\code{"integer"} equal to the number of partitions.}
\item{list("type")}{Object of class \code{"character"} defining type of
data.}\item{:}{Object of class \code{"character"} defining type of data.} }
}

\examples{

#none

}
\seealso{
End-point array digital PCR: \code{\linkS4class{adpcr}}.

Droplet digital PCR: \code{\linkS4class{dpcr}}.
}
\author{
Michal Burdukiewicz.
}
\keyword{classes}
\keyword{real-time}
