% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show_dpcr.R
\docType{methods}
\name{show-methods}
\alias{show}
\alias{show,adpcr-method}
\alias{show,ddpcr-method}
\alias{show-methods}
\title{Methods for Function \code{show}}
\arguments{
\item{object}{an object of class \code{\linkS4class{adpcr}} or
\code{\linkS4class{ddpcr}}.}
}
\description{
Expands function \code{\link[methods]{show}} allowing showing objects of the
class \code{\linkS4class{adpcr}} to or \code{\linkS4class{ddpcr}}.
}
\examples{
#array dpcr
ptest <- sim_adpcr(400, 765, 5, FALSE, n_panels = 1)
show(ptest)

#multiple experiments
ptest <- sim_adpcr(400, 765, 5, FALSE, n_panels = 5)
show(ptest)

#droplet dpcr - fluorescence
dropletf <- sim_ddpcr(7, 20, times = 5, fluo = list(0.1, 0))
show(dropletf)

#droplet dpcr - number of molecules
droplet <- sim_ddpcr(7, 20, times = 5)
show(droplet)
}
\author{
Michal Burdukiewicz.
}
\keyword{methods}
\keyword{utilities}

